+++ CHANGELOG.md
@@ -11,8 +11,8 @@ This project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html
 
 ### New Features
 
-- Added support for introspecting a scalars `@specifiedBy` field on servers
-  supporting GraphQL 2021
+- Added support for introspecting the `specifiedByUrl` field of scalars on
+  servers supporting GraphQL 2021
 - `cynic-introspection` can now output SDL
 - Added `CapabilitiesQuery` to `cynic-introspection`, which can detect
   which version of the GraphQL specification a server supports.